project1_simple_chatbot/requirements.txt
# 项目1：简单聊天机器人 - 依赖包
# Core dependencies
openai>=1.3.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: for enhanced logging and configuration
pydantic>=2.0.0
colorama>=0.4.6